+++ src/algo/hash.h
@@ -10,6 +10,8 @@
  */
 #pragma once
 
+#include <cstdint>
+
 /**
  * @brief Simple Hash functions.
  *
@@ -21,29 +23,35 @@ public:
     /**
      * @brief Famous fast 32 bits hash function used by the sdbm public-domain database library.
      *
+     *  Also know by its multiplier "x65599". Originates from the "Dragon Book" (1986).
+     *
      * @see http://stackoverflow.com/a/14409947/1163698
+     * @see http://www.cse.yorku.ca/~oz/hash.html
      */
-    static unsigned long sdbm(const char* apStr);
+    static uint32_t sdbm(const char* apStr);
 
     /**
      * @brief Famous very fast 32 bits hash function by Daniel J. Bernstein.
      *
+     * @warning It has some flaws for hashing binary data [0:255] because it uses the "33" multiplier (not high enough)
+     *
      * @see http://stackoverflow.com/a/14409947/1163698
+     * @see http://www.cse.yorku.ca/~oz/hash.html
      */
-    static unsigned long djb2(const char* apStr);
+    static uint32_t djb2(const char* apStr);
 
     /**
      * @brief Famous 32 bits FNV-1a hash function by Glenn Fowler, Landon Curt Noll, and Phong Vo.
      *
      * @see http://www.isthe.com/chongo/tech/comp/fnv/index.html#FNV-1a
      */
-    static unsigned long fnv1a32(const char* apStr);
+    static uint32_t fnv1a32(const char* apStr);
 
     /**
      * @brief Famous 64 bits FNV-1a hash function by Glenn Fowler, Landon Curt Noll, and Phong Vo.
      *
      * @see http://www.isthe.com/chongo/tech/comp/fnv/index.html#FNV-1a
      */
-    static unsigned long long fnv1a64(const char* apStr);
+    static uint64_t fnv1a64(const char* apStr);
 };
